ExportPipeline: Don't call waitUntilFinished() on template generation jobs
Template generation shouldn't call waitUntilFinished() on the pipeline.  Since
a dataflow job is not created, it will raise an exception trying to wait on a
non-existent job. If template location is null, it's a "regular" dataflow job
and can be waited for.

This is identical to the recent ImportPipeline change.

PiperOrigin-RevId: 218391539