Update PubSubToBigQuery to use subscription input instead of topic.
The topic input makes the workflow non-deterministic because there is no
guarantee of when the workflow will create the subscription for the
provided topic. Thus, messages sent to the topic could be dropped.
Requiring an existing subscription guarantees that any unread messages
published to the topic will be read eventually.