Added a validation for non-existing tables in the DB.
Also removed the CreateIndexes step (and the flag), as for importing text files, we assume the tables have already been created. And when loading large amount of data, it's recommended to drop secondary indexes and recreate them after the loading is done. It's not a step the pipeline is supposed to do (we don't know the schema of the indexes anyway).

PiperOrigin-RevId: 238518371