Update TableRow conversion to support dates
The TableRow object in BigQuery requires the dates columns to be formatted before inserting into BigQuery. With the data set with date fields, the template is breaking because Jdbc convertor is not taking date fields into account. This change is to modify the ResultSetToTableRow method to format the date fields in the format acceptable to BigQuery.