Modify default grouping for SpannerIO.Write to 100,000 cells and 100MB
The grouping is so that the rows can be sorted and batched for efficient spanner writes.

The default grouping is for 1000*batch_size which works out to a max of 5,000,000 mutated cells and 1GB of mutated values.

This can OOM if the rows are very narrow and small because the overhead of storing a mutated row is significantly larger than the number of bytes mutated.

This CL modifies reduces the default grouping factor to 100 and max cells per batch to 1000.

PiperOrigin-RevId: 275460692