Move "Wait For Previous Depth" to before avro files are read.
Currently, all avro file read stages are runnable at the start of the pipeline, but the pipeline waits for the mutations previous depth to complete before the rows read from avro to be written to spanner.
This can lead to child table avro files being read too 'early', and then the results having to wait for a significant amount of time for the previous write step to complete, which may use up resources.

This change modifies the pipeline such that the avro files are only read _after_ the previous depth has completed.

PiperOrigin-RevId: 275475024