Disable spilling in Beam's AvroIO.Write
By default, Beam's WriteFiles transform only writes to 20 filenames per Window (which in the case of the ExportTransform is the Global Window).

Other elements are pushed into a side pipeline to be processed when the input is completely read. This then does a GroupByKey, potentially on many TBs of records.

Disabling spilling simplifies and speeds up the pipeline by about 40% on a test DB, potentially more on larger DBs

PiperOrigin-RevId: 289174197