Cloud Spanner Avro import export for stored generated columns.
Export: The values of a stored generated column are not exported as they are semantically logical entities. The definition of a stored generated column is exported to the Avro schema as a null-type record field with the column definition being custom properties of the field. Generated columns in backfilling are ignored as if they don't exist in the schema.

Import: The generated column definition in the Avro schema will be used to recreate the generated column. Spanner will compute the generate column values automatically during import.
PiperOrigin-RevId: 325167899