Cloud Spanner CSV import export for stored generated columns.
Export: The values of a stored generated column are not exported as they are semantically logical entities.

Import: The users must pre-create the tables in Spanner and provide a json manifest. If the table has generated columns, the manifest must include an explicit list of non-generated columns to import. Spanner will compute the generated column values automatically during import.
PiperOrigin-RevId: 325523996