Fix handling nulls when converting from JSONs
The CL updates how the JSON files are being converted to GenericRecord instances.
The original method that used the built in Avro converter didn't handle nullable
value from Dataplex json files. (The builtin JSON decoder is expecting the nullable values that are not null to be "annotated" with a type, while Dataplex's JSONs, as expected, just contain the values.)

PiperOrigin-RevId: 406010522