Dataplex Tiering template: fix for the handling of the TIME fields.
The CL updates the BigQuery to GCS transform to handle the fields with the type TIME. There is misalignment between Apache Beam and BigQuery API when translating
the TIME BigQuery type to Avro type. Beam translates it as `string` while BigQuery translates it as `int` with a "logicalType" `time-micros`.

Also the TAP tests were timing out sometimes, successful runs were close to 900 seconds limit for large tests.

PiperOrigin-RevId: 440182344