Set maxElapsedMillis on HttpEventPublisher to speedup Splunk tests
I noticed that a fair chunk of the unit test time is spent on a single Splunk test case (`HttpEventPublisherTest - unrecognizedSelfSignedCertificateTest`).

It expects an `SSLHandshakeException` for an invalid/self-signed certificate, which is correct, but the default configurations for `HttpEventPublisher`/`HttpRequest` is to use an exponential backoff strategy for any subclass of `IOException` (note that `SSLHandshakeException -> SSLException -> IOException`).

PiperOrigin-RevId: 468677061