The template will now ingest a HarbourBridge generated session file and use it to modify the change events later on.
This specific change allows us to read the session file from GCS and store it into memory. We add classes to represent the session file structure which we populate from the json file using the GSON library.

We also want to maintain backwards compatibility by allowing omission of the session file parameter. If the session file path parameter is not specified, the template behaves exactly like it used to earlier i.e. perform no modifications to the change event.

PiperOrigin-RevId: 471599035