Fix a bug in JdbcConverters that caused timestamps to have incorrect microseconds portion.
This affects JdbcToBigQuery and DataplexJdbcIngestion templates and fields with TIMESTAMP SQL type in the source database.
For example, with the JdbcToBigQuery template, a timestamp 01:02:03.555777 from the source SQL database used to be saved in BigQuery as 01:02:03.000555.

Also switched from SimpleDateFormatter to DateTimeFormatter in JdbcConverters, as the former is not thread-safe.

PiperOrigin-RevId: 505186670