Spanner Vertex AI Integration: Add support for required columns.
Since these are not deployed in production yet, we have to ensure that
the pipeline will work before and after this happens. This can be
achieved by simply not printing default required=TRUE value.

Before the release, all models will have all columns as required. The
pipeline will skip printing this option, avoiding unsupported error in
the parser.

After the release, users will be able to set required to FALSE, in which
case the pipeline will emit the DDL and parser will accept it.

Later, we can remove this crutch completely.