Fix support of multiple key definitions for edge source/target nodes
This only concerns the "transposed" spec syntax.

If the keys for the source/target node of the edge were encoded as
a JSON object, the existing implementation would only pick the
"first" key and define a mapping only for that one.

JSON objects are deserialized into Java's HashMap instances, and
these do not provide any guarantees for the iteration order, which
makes the key definition very difficult to reason about.

Moreover, the "verbose" syntax allows to define multiple keys and
these are properly mapped to multiple mappings.

This commit aligns the behavior between the two syntaxes' mappers.