Make source iteration order deterministic
Note: this is only a quick fix, as a proper rework of target and
action processing needs to happen, since there can be dependencies
between targets and targets, between actions and targets as well
as targets and actions.

Turns out this is the graph problem... but the template does not
solve it at all today and may fail with NPEs.