Fix custom query flaky test
This introduces two node targets, in order to deduplicate the IDs
used in the tests at the source query level (with GROUP BY).
Doing so eliminates the concurrent MERGE problem, that fails hard
when uniqueness constraints are in place (which is the case for
this test's spec).

The other test has also been deleted since it overlaps now much
more with what the fixed test covers.