Bug template: ask template version instead of Beam version
There's not a lot of point in asking for Beam version, and this list gets stale frequently. If we know what's the version of template, we are able to easily figure out the Beam version.
Also, the template version might tell us more about the code that is being executed, since we have about 10 template releases per Beam version.