Enfore csvFormat and delimiter as required parameters for GCS_CSV_to_BigQuery template
Changing csvFormat to be required so that it's clear to user which format they're using.

Change delimiter to be required due to bug https://github.com/apache/beam/issues/30261
that causes we won't be able to get the value of delimiter dynamically. This change is
fine since delimiter is a common paramter in CSV and lots of users set it anyway.

Signed-off-by: keyliug <keyliu@google.com>