[Cloud Spanner to BigQuery Template] Fix a NPE when one of the source fields contains a NULL value.
Without this fix, the template expects all the source record fields to have non-NULL values. If there is one, an exception is thrown:

org.apache.beam.sdk.util.UserCodeException:java.lang.NullPointerException: Column field0 contains NULL value.

The reason is that SpannerUtils class tries to call struct.getString/getLong/etc. unconditionally, and these methods throw an exception if the value is null. The fix is to skip such fields and not include them in the JSON object (TableRow) being written to BigQuery.
This will also allow a target field in BigQuery to not exist at all if the corresponding source field in Spanner is always null.