feat: add orderBy column and order
When multiple changes are done to the same rows, the pipeline
assumes that the first seen row is the first updated one, and the
last will contain the up to date record. Howevever, Avro writers,
and grouping logic does not usually guarantee order. To solve this,
the pipeline supports one column name by which the data with the
same primary key will be sorted.