refactor: add missing helper methods in CypheGeneratorTest
Observed: tons of repetitive testing setups for import statements, just to assert a string.
And schema statement tests have helper method to aid in setting up assertions for schema
statments

This change add a similar helper for import statements, in a base class for de-clutter.
The goal is to keep all hard testing data in the original test scope, e.g. expected
strings or unique setups.