feat: enforce cypher 5 for generated queries
Neo4j version 2026.0 introduces the possibility to specify if cypher version 5 or 25 should be
used. In order to maximise backwards compatibility and minimize regressions we want to use
cypher version 5 in case of generated queries when possible.

This change extends Neo4jCapabilities to properly flag for a version that ships with cypher
version statements.

Then also extends Neo4jRowWriterTransform to check if "CYPHER 5" can be safely prepended.