feat: enforce cypher 5 for generated queries (#2620)
* refactor: add missing helper methods in CypheGeneratorTest

Observed: tons of repetitive testing setups for import statements, just to assert a string.
And schema statement tests have helper method to aid in setting up assertions for schema
statments

This change add a similar helper for import statements, in a base class for de-clutter.
The goal is to keep all hard testing data in the original test scope, e.g. expected
strings or unique setups.

* feat: enforce cypher 5 for generated queries

Neo4j version 2026.0 introduces the possibility to specify if cypher version 5 or 25 should be
used. In order to maximise backwards compatibility and minimize regressions we want to use
cypher version 5 in case of generated queries when possible.

This change extends Neo4jCapabilities to properly flag for a version that ships with cypher
version statements.

Then also extends Neo4jRowWriterTransform to check if "CYPHER 5" can be safely prepended.

* test: add mocked tests for capabilities call

Trying to make Codecov happy, this change introduces two new tests for makings sure we dial a
database when calling capabilities()

* fix: updated neo4j cypher 5 capability version

selecting cypher 5 is available since neo4j version 5.21, a bit earlier than before's 2025.06
this change updates from 2025.06 to 5.21

this change also contain test fixes where order of expected values are determined in the test query.

Co-authored-by: Ali Ince <ali-ince@users.noreply.github.com>
Co-authored-by: Eugene Rubanov <eugene.rubanov@neo4j.com>

---------

Co-authored-by: Ali Ince <ali-ince@users.noreply.github.com>
Co-authored-by: Eugene Rubanov <eugene.rubanov@neo4j.com>