Migrate from gsutil to gcloud storage (#2721)
* Update cloudbuild.yaml

This PR is part of the on going effort to migrate from the legacy gsutil tool to the new and improved gcloud storage command-line interface. gcloud storage is the recommended and modern tool for interacting with Google Cloud Storage, offering better performance, unified authentication, and a more consistent command structure with other gcloud components. 🚀

* Update cloudbuild.yaml

revert changes