Add check on max STRING length and also run for PG dialect
Note, this also updates the length of the BYTES row that we migrate because the previous value did not actually exceed
the limit (it ended up being 7MiB, less than the 10MiB max). This now writes a value that's actually larger than the
limit and verifies that the row is NOT saved.