[datastream-to-spanner] Unable to convert field timestamp to long (#3035)
[b/458064173](https://b.corp.google.com/issues/458064173)

#### Problem
We create some new columns to hold metadata in the shadow tables that are created. These shadow tables also have the Primary Key columns from the data table. When the data table has columns of the same name as the metadata columns we create, it gets overwritten because of which we get data type conversion errors. For eg. a primary key column "timestamp TIMESTAMP" gets overwritten to "timestamp INT64".

#### Fix
To ensure that we don't have these kind of clashes, we now dynamically compute the name of the shadow column by checking it against the list of primary key column names. In case a column with the same name is found, we add "shadow_" as a prefix iteratively till we get a unique column name for the shadow table. 

### Testing:
- Built container with changes: gs://ea-functional-tests/templates-aastha-2025-12-09/flex/Cloud_Datastream_to_Spanner
-  Confirmed in DDL of Shadow table that original "timestamp" column of type TIMESTAMP is present along with new columns: shadow_timestamp, log_file and log_position
- Ran a live migration job with above template and ran INSERT, UPDATE and DELETE statements containing "timestamp TIMESTAMP" column as Primary Key in data table - and these passed without errors