Add integration tests to verify table and index limits for bulk spanner migration from PostgreSQL (#3103)
* Add check on max STRING length and also run for PG dialect

Note, this also updates the length of the BYTES row that we migrate because the previous value did not actually exceed
the limit (it ended up being 7MiB, less than the 10MiB max). This now writes a value that's actually larger than the
limit and verifies that the row is NOT saved.

* Add check on max row size for both dialects

Currently ignored because it causes OOMs on low-provisioned VMs/DBs

* Fix spotless

* Add integration tests to check string and cell max size limits when migrating from Postgres

* Add integration tests to check primary key and index max size limits when migrating from Postgres

* Add integration tests to check row max size limits when migrating from Postgres